Add support to generate hardware configuration
Added serializers for different format
Refactored the build commands to return an index for serialization

+++ src/resolver.rs
@@ -1621,7 +1621,7 @@ fn register_string_type(index: &mut Index, is_wide: bool, len: usize) -> String 
                 } else {
                     StringEncoding::Utf8
                 },
-                size: typesystem::TypeSize::LiteralInteger(len as u32 + 1),
+                size: typesystem::TypeSize::LiteralInteger(len as i64 + 1),
             },
         });
     }
